KalmanFilterToolbox

Documentation for KalmanFilterToolbox.

KalmanFilterToolbox.IWP

KalmanFilterToolbox.discretize

KalmanFilterToolbox.ekf_update

KalmanFilterToolbox.get_backward_transition

KalmanFilterToolbox.linearize

KalmanFilterToolbox.predict

KalmanFilterToolbox.projectionmatrix

KalmanFilterToolbox.smooth

KalmanFilterToolbox.sqrt_predict

KalmanFilterToolbox.sqrt_update

KalmanFilterToolbox.update

KalmanFilterToolbox.IWP — Type

IWP(wiener_process_dimension::Integer, num_derivatives::Integer)

Integrated Wiener Process.

By itself it does not have much utility right now, but together with discretize it provides discrete transition matrices that are useful for defining discrete state-space models.

KalmanFilterToolbox.discretize — Method

discretize(iwp::IWP, dt::Real)

Discretize the integrated Wiener process.

Computes the discrete transition matrices for a time step of size dt.

KalmanFilterToolbox.ekf_update — Method

ekf_update(m, C, y, h, R)

Update / correct the state based on a nonlinear observation.

This function does two things:

it linearizes the observation function h at the mean m, with linearize, and

it calls update to perform an update on the now-linear model.

KalmanFilterToolbox.get_backward_transition — Method

get_backward_transition(m, C, mpred, Cpred, A)

Compute the affine backward transition model used for smoothing.

Returns parameters for a transition $x_n^S \mid x_{n+1}^S \sim \mathcal{N}(G x_{n+1}^S + b, Λ)$, computed with

\[\begin{aligned} G &= C * A^\top C_p^{-1} \\ b &= m - G m_p \\ Λ &= C - G C_p G^\top. \end{aligned}\]

To smooth, just predict backwards.

KalmanFilterToolbox.linearize — Method

linearize(h::Function, m::AbstractVector)

Linearize the nonlinear function h at the location m.

Approximate h with $h(x) \approx H x + b$, where

\[\begin{aligned} H &= J_h(m), \\ b &= h(m) - H * m. \end{aligned}\]

The Jacobian is computed with automatic differentiation via ForwardDiff.jl.

KalmanFilterToolbox.predict — Method

predict(m, C, A, b, Q)

Predict the next state estimate for an affine Gaussian transition.

Given an affine transition model $x_{n+1} \mid x_n \sim \mathcal{N}(x_{n+1}; A x_n + b, Q)$ and a current state estimate $x_n \sim \mathcal{N} (m, C)$, predict computes the prediction estimate

\[\begin{aligned} x_{n+1} \sim \mathcal{N} (A m + b, A C A^\top + Q). \end{aligned}\]

KalmanFilterToolbox.projectionmatrix — Method

projectionmatrix(iwp::IWP, derivative::Integer)

Compute the projection matrix that maps the state to the specified derivative.

KalmanFilterToolbox.smooth — Method

smooth(mcurr, Ccurr, mnext, Cnext, A, Q)

Smoothing step to update the "current" state estimate on the "next" one.

This implementation internally calls predict to (re-)compute the prediction estimate, and then perform the backwards smoothing.

In most cases, pre-computing the backwards transitions directly via get_backward_transition and then just predicting backwards might be the preferred strategy.

KalmanFilterToolbox.sqrt_predict — Method

sqrt_predict(m, CL, A, b, QL)

Predict the next state estimate for an affine Gaussian transition, in square-root form.

In principle, this function does the same as predict, but the covariances are given as and returned in square-root form. That is, the equivalent call to predict(m, C, A, b, Q) would be sqrt_predict(m, CL, A, b, QL), where C = CL * CL' and Q = QL * QL'.

KalmanFilterToolbox.sqrt_update — Method

sqrt_update(m, CL, y, H, b, RL)

Update / correct the state based on the affine observation, in square-root form.

In principle, this function does the same as update, but the covariances are given as and returned in square-root form. That is, the equivalent call to update(m, C, y, H, b, R) would be sqrt_update(m, CL, y, H, b, RL) where C = CL * CL' and R = RL * RL'.

KalmanFilterToolbox.update — Method

update(m, C, y, H, b, R)

Update / correct the state based on the affine observation.

Given a Gaussian $x \sim \mathcal{N}(m, C)$ an affine observation model

\[\begin{aligned} y \mid x \sim \mathcal{N}(y; H x + b, R), \end{aligned}\]

and a data point $y$, update computes the posterior $x \mid y$.