Gauss-Markov Processes

Things that relate to continuous-time Gauss-Markov processes and their discretization.

Processes

KalmanFilterToolbox.IWP — Type

IWP(wiener_process_dimension::Integer, num_derivatives::Integer)

Integrated Wiener process.

By itself it does not have much utility right now, but together with discretize it provides discrete transition matrices that are useful for defining discrete state-space models.

KalmanFilterToolbox.IOUP — Type

IOUP(wiener_process_dimension::Integer, num_derivatives::Integer, drift_speed::Real)

Integrated Ornstein-Uhlenbeck process.

KalmanFilterToolbox.Matern — Type

Matern(wiener_process_dimension::Integer, num_derivatives::Integer, lengthscale::Real)

Matern process. Typically, they are defined with a half-integer parameter $ν$. To get the corresponding process here, set num_derivatives = ν+1/2.

Utilities

KalmanFilterToolbox.discretize — Function

discretize(p::AbstractGaussMarkovProcess, dt::Real)

Discretize the Gauss-Markov process.

Computes the discrete transition matrices for a time step of size dt.

discretize(sde::LTISDE, dt::Real)

Compute the discrete transition via the matrix fraction decomposition.

See also: matrix_fraction_decomposition

KalmanFilterToolbox.projectionmatrix — Function

projectionmatrix(process::AbstractGaussMarkovProcess, derivative::Integer)

Compute the projection matrix that maps the state to the specified derivative.

KalmanFilterToolbox.preconditioner — Function

preconditioner(p::AbstractGaussMarkovProcess, dt::Real)

Computes a preconditioner, which is helpful for numerically stable implementation.

This corresponds to the preconditioner as described by [1], developed for integrated Wiener process priors (IWP), but it might also be useful for other priors.

[1] Krämer & Hennig, "Stable Implementation of Probabilistic ODE Solvers", 2020.

KalmanFilterToolbox.matrix_fraction_decomposition — Function

matrix_fraction_decomposition(drift::AbstractMatrix, dispersion::AbstractVecOrMat, dt::Real)

KalmanFilterToolbox.preconditioned_discretize — Function

preconditioned_discretize(p::IWP)

Preconditioned discretization of the integrated Wiener process.

For the IWP, the preconditioned discretization is independend of the current step size and known in closed form. This is very helpful for numerically stable implementations and high efficiency.

SDE and tools

KalmanFilterToolbox.LTISDE — Type

LTISDE(drift::AbstractMatrix, dispersion::AbstractVecOrMat)

Linear time-invariant SDE

\[\begin{aligned} dx = F x dt + L dβ \end{aligned}\]

with drift F and dispersion L.

KalmanFilterToolbox.to_1d_sde — Function

to_1d_sde(p::AbstractGaussMarkovProcess)

Build a LTISDE that corresponds to a single dimension of the given Gauss-Markov process.

Just working on a single dimension leads to more efficient code, so that the full matrices can be built at the very end (e.g. after computing the one-dimensional transition matrices via matrix fraction decomposition) with a Kronecker product.